Corrected an issue where integer overflow would cause a false positive for sparse matrix reserved space being too large.

+++ source/cdouble/matrices/sparse_column_compressed.c
@@ -109,7 +109,7 @@ jmtx_result jmtxz_matrix_ccs_new(
     if (reserved_entries == 0)
     {
         reserved_entries = DEFAULT_RESERVED_ELEMENTS;
-        reserved_entries = reserved_entries < cols * rows ? reserved_entries : cols * rows;
+        reserved_entries = reserved_entries < ((uint64_t)cols * (uint64_t)rows) ? reserved_entries : ((uint64_t)cols * (uint64_t)rows);
     }
 
     uint32_t* offsets = NULL;
@@ -178,7 +178,7 @@ jmtx_result jmtxzs_matrix_ccs_new(
     {
         return JMTX_RESULT_BAD_PARAM;
     }
-    if (reserved_entries > cols * rows)
+    if (reserved_entries > ((uint64_t)cols * (uint64_t)rows))
     {
         return JMTX_RESULT_BAD_PARAM;
     }
